+++ packages/powersync/lib/src/database/powersync_database.dart
@@ -364,11 +364,26 @@ abstract base class PowerSyncDatabase extends SqliteConnection {
   /// would be empty.
   ///
   /// To preserve data in local-only tables, set [clearLocal] to false.
-  Future<void> disconnectAndClear({bool clearLocal = true}) async {
+  ///
+  /// A [soft] clear deletes publicly visible tables, but keeps internal copies
+  /// of data synced in the database. This usually means that if [connect] is
+  /// later called again with a JWT from the same user, the first sync is very
+  /// fast because all internal data is still available. When a different user
+  /// logs in, no old data would be visible at any point. Using soft deletes
+  /// is recommended where it's not a security issue that old data could be
+  /// reconstructed from internal database tables.
+  Future<void> disconnectAndClear({
+    bool clearLocal = true,
+    bool soft = false,
+  }) async {
     await disconnect();
 
     await writeTransaction((tx) async {
-      await tx.execute('select powersync_clear(?)', [clearLocal ? 1 : 0]);
+      var flags = 0;
+      if (clearLocal) flags |= 1;
+      if (soft) flags |= 2;
+
+      await tx.execute('select powersync_clear(?)', [flags]);
     });
     await _connections.resolveOfflineSyncStatus();
   }